fix series gql
