+++ src/documents/works/index.ts
@@ -8,20 +8,6 @@ export const workDocument = graphql(`
   }
 `);
 
-export const seriesDocument = graphql(`
-  query GetSeries($series_id: String!) {
-    works(
-      where: { series_id: { _eq: $series_id } }
-      order_by: [{ has_episodes: desc }]
-    ) {
-      ...FragmentWork
-      episodes(order_by: { number: desc_nulls_last }, limit: 8) {
-        ...FragmentEpisode
-      }
-    }
-  }
-`);
-
 export const searchWorksDocument = graphql(`
   query SearchWorks($search: String!, $limit: Int) {
     search_works(
